pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "concatenator"
version = "0.1.0"
description = "Musical concatenation tools and MIDI dataset utilities"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Nathan Turczan"}
]
keywords = ["music", "midi", "concatenation", "music21", "lilypond", "composition"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Multimedia :: Sound/Audio :: MIDI",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "music21>=9.0.0",
    "numpy>=1.20.0",
    "graphviz>=0.20.0",
]

[project.optional-dependencies]
audio = [
    "pydub>=0.25.0",
    "pretty_midi>=0.2.10",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
all = [
    "pydub>=0.25.0",
    "pretty_midi>=0.2.10",
]

[project.scripts]
concatenate-midi = "concatenator.scripts.concatenate_midi:main"
concatenate-audio = "concatenator.scripts.concatenate_audio:main"

[project.urls]
Homepage = "https://github.com/nathanturczan/concatenator"
Repository = "https://github.com/nathanturczan/concatenator"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"
